pyproject.toml
[tool.poetry]
name = "gitportfolio"
version = "0.1.0"
description = "GitHub action to generate a list of repositories"
keywords = [
    "python",
    "github",
    "readme",
]
authors = ["George-Andrei Iosif <hi@iosifache.me>"]
license = "MIT"
repository = "https://github.com/iosifache/gitportfolio"
homepage = "https://project.com"
readme = "README.pypi.md"
classifiers = [
    "Environment :: Console",
    "Operating System :: OS Independent",
]
include = [
    "LICENSE",
]
packages = [{include = "gitportfolio"}]

[tool.poetry.urls]
"Changelog" = "https://github.com/iosifache/gitportfolio/CHANGELOG.md"
"Issue Tracker" = "https://github.com/iosifache/gitportfolio/issues"
"LinkedIn" = ""
"Twitter" = ""

[tool.poetry.dependencies]
python = "^3.11"
pygithub = "^2.1.1"
tabulate = "^0.9.0"
pyyaml = "^6.0.1"
click = "^8.1.7"
rich = "^13.7.0"

[tool.poetry.scripts]
gitportfolio = "gitportfolio.main:main"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
mypy = "^1.0.0"
ruff = "^0.0.246"
coverage = {extras = ["toml"], version = "^7.1.0"}
poethepoet = "^0.18.1"
pytest = "^7.2.1"
types-pyyaml = "^6.0.12.12"
types-tabulate = "^0.9.0.3"

[tool.black]
line-length = 79
preview = true
include = '\.pyi?$'
extend-exclude = """
(
    ^/.venv/*
    ^/dist/*
)
"""

[tool.ruff]
src = ["gitportfolio", "tests"]
select=[
    # Pyflakes
    "F",
    # pycodestyle
    "E",
    "W",
    # mccabe
    "C90",
    # isort
    "I",
    # pep8-naming
    "N",
    # pyupgrade
    "UP",
    # flake8-2020
    "YTT",
    # flake8-annotations
    "ANN",
    # flake8-bandit
    "S",
    # flake8-blind-except
    "BLE",
    # flake8-boolean-trap
    "FBT",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-commas
    "COM",
    # flake8-comprehensions
    "C4",
    # flake8-datetimez
    "DTZ",
    # flake8-debugger
    "T10",
    # flake8-django
    "DJ",
    # flake8-errmsg
    "EM",
    # flake8-executable
    "EXE",
    # flake8-implicit-str-concat
    "ISC",
    # flake8-import-conventions
    "ICN",
    # flake8-logging-format
    "G",
    # flake8-no-pep420
    "INP",
    # flake8-pie
    "PIE",
    # flake8-print
    "T20",
    # flake8-pyi
    "PYI",
    # flake8-pytest-style
    "PT",
    # flake8-quotes
    "Q",
    # flake8-return
    "RET",
    # flake8-simplify
    "SIM",
    # flake8-tidy-imports
    "TID",
    # flake8-type-checking
    "TCH",
    # flake8-unused-arguments
    "ARG",
    # flake8-use-pathlib
    "PTH",
    # eradicate
    "ERA",
    # pandas-vet
    "PD",
    # pygrep-hooks
    "PGH",
    # Pylint
    "PL",
    # tryceratops
    "TRY",
    # flake8-raise
    "RSE",
    # flake8-self
    "SLF",
    # Ruff-specific rules
    "RUF",
]
ignore = [
    "ANN101"
]
line-length = 79

[tool.ruff.isort]
known-first-party= ["gitportfolio"]
known-local-folder = ["gitportfolio"]

[tool.ruff.pydocstyle]
convention = "google"

[tool.mypy]
files = [
  "gitportfolio",
  "scripts",
  "tests"
]
show_error_codes = true
namespace_packages = true
explicit_package_bases = true
ignore_missing_imports = true

[tool.coverage.run]
omit = [
    "tests/*"
]

[tool.poe.tasks]

    [tool.poe.tasks.test]
    help = "Run all unit tests."
    cmd  = "coverage run -m pytest -vv tests/"

    [tool.poe.tasks.coverage]
    help = "Report the coverage"
    cmd = "coverage report -m"

    [tool.poe.tasks.covtest]
    help = "Run the unit tests and report the coverage."
    sequence = ["test", "coverage"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
